Spectra MS backend to read data from files in mzPeak format

The MsBackendMzPeak is an implementation of the Spectra::MsBackend() class for Spectra::Spectra() objects adding support for MS data import from the HUPO-PSI mzPeak file format. MsBackendMzPeak is an on-disk implementation keeping only minimal information in memory while loading any data (peaks data or spectra metadata) from the original mzPeak files on-the-fly.

MsBackendMzPeak()

# S4 method for class 'MsBackendMzPeak'
show(object)

Arguments

object

An MsBackendMzPeak object.

Details

The MsBackendMzPeak backend uses functionality from the mzpeak R package for accessing data in mzPeak files.

The MsBackendMzPeak extends the Spectra::MsBackendCached()backend allowing thus to intermediately store spectrum metadata (*spectra variables*) in adata.frame` within the object (i.e., in memory).

Creation of backend objects

New MsBackendMzPeak backend objects can be created with the MsBackendMzPeak() function.

Subset, merge and filter data

Subset...

Data access

Data access functionality.

Author

Johannes Rainer, Gabriele Tomè